+++ NetVis/Configuration/OwnershipSettings.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using System;
+using UnityEngine;
 using Unity.Multiplayer.Tools.Adapters;
 using Unity.Multiplayer.Tools.Common.Visualization;
 
@@ -8,22 +9,64 @@ namespace Unity.Multiplayer.Tools.NetVis.Configuration
     {
         public bool MeshShadingEnabled { get; set; } = true;
         public bool TextOverlayEnabled { get; set; } = true;
+        public event Action ColorsChanged;
 
-        public Color ServerHostColor { get; } = CategoricalColorPalette.GetColor(0);
+#if UNITY_EDITOR
+        public OwnershipSettings()
+        {
+            CustomColorSettings.DataChanged += () => ColorsChanged?.Invoke();
+        }
+#endif
 
-        // We use a dictionary here, for two reasons
-        // 1. Connected clients can come out of order
-        // 2. We want default settings, but there's no guarantee clientIds use sequential integers
-        internal ReadonlyClientColorDictionary ClientColors { get; } = new();
-    }
+        public Color ServerHostColor
+        {
+            get
+            {
+#if UNITY_EDITOR
+                if (CustomColorSettings.HasColor(0))
+                {
+                    return CustomColorSettings.GetColor(0);
+                }
+#endif
 
-    /// <remarks>
-    /// Temporary, readonly stand-in for a dictionary of ClientId -> Color, until we reintroduce customization
-    /// of client colors in a future release. 2023-04-06
-    /// </remarks>>
-    class ReadonlyClientColorDictionary
-    {
-        public bool TryGetValue(ClientId clientId, out Color color) =>
-            CategoricalColorPalette.TryGetColor((int)clientId, out color);
+                return CategoricalColorPalette.GetColor(0);
+            }
+        }
+
+        internal Color GetClientColor(ClientId clientId)
+        {
+#if UNITY_EDITOR
+            if (CustomColorSettings.HasColor((int)clientId))
+            {
+                return CustomColorSettings.GetColor((int)clientId);
+            }
+#endif
+
+            return CategoricalColorPalette.GetColor((int)clientId);
+        }
+
+        internal void SetCustomColor(ClientId clientId, Color color)
+        {
+#if UNITY_EDITOR
+            CustomColorSettings.SetColor((int)clientId, color);
+            ColorsChanged?.Invoke();
+#endif
+        }
+
+        internal void ResetCustomColor(ClientId clientId)
+        {
+#if UNITY_EDITOR
+            CustomColorSettings.RemoveColor((int)clientId);
+            ColorsChanged?.Invoke();
+#endif
+        }
+
+        internal void ResetCustomColors()
+        {
+#if UNITY_EDITOR
+            CustomColorSettings.ClearColors();
+            ColorsChanged?.Invoke();
+#endif
+        }
     }
 }